Hesse/liitetiedostot kalenterimerkintoihin

+++ crm/kalenteri.php
@@ -68,10 +68,10 @@ $vertaa     = "'".implode("','", $ruksatut)."'";	// tehd‰‰n mysql:n ymm‰rt‰m‰ mu
 
 if (trim($konserni) != '') {
 	$query = "SELECT distinct yhtio FROM yhtio WHERE (konserni = '$yhtiorow[konserni]' and konserni != '') or (yhtio = '$yhtiorow[yhtio]')";
-	$result = mysql_query($query) or pupe_error($query);
+	$result = pupe_query($query);
 	$konsernit = "";
 
-	while ($row = mysql_fetch_array($result)) {
+	while ($row = mysql_fetch_assoc($result)) {
 		$konsernit .= " '".$row["yhtio"]."' ,";
 	}
 	$konsernit = " and kalenteri.yhtio in (".substr($konsernit, 0, -1).") ";
@@ -88,13 +88,25 @@ $query = "	SELECT yhtio
 			and kuka	= '$kukarow[kuka]'
 			and nimi like '%asiakasmemo.php'
 			and alanimi = ''";
-$result = mysql_query($query) or pupe_error($query);
+$result = pupe_query($query);
 
 if (mysql_num_rows($result) > 0) {
 	$asmemolinkki = TRUE;
 }
 else {
 	$asmemolinkki = FALSE;
+}
+
+// Poistetaan liitetiedosto kalenterimerkinn‰st‰
+// mik‰li GET-parametreist‰ lˆytyy poista_liite
+if( isset($poista_liite) and (int) $poista_liite > 0 ) {
+	$poista_liite = (int) $poista_liite;
+
+	$query = "	DELETE FROM liitetiedostot
+				WHERE tunnus = $poista_liite
+				AND liitos = 'kalenterimerkint‰'
+				AND yhtio = '$kukarow[yhtio]'";
+	pupe_query($query);
 }
 
 echo "<font class='head'>".t("Kalenteri")."</font><hr>";
@@ -172,12 +184,14 @@ if ($tee == 'LISAA') {
 	}
 
 	if ($ok == "OK") {
-		if ($tunnus != '') {
+		$tunnus = (int) $tunnus;
+
+		if ($tunnus > 0) {
 			$query = "	DELETE
 						FROM kalenteri
-						WHERE tunnus='$tunnus'
+						WHERE tunnus=$tunnus
 						$konsernit";
-			mysql_query($query) or pupe_error($query);
+			pupe_query($query);
 		}
 
 		if ($kenelle == "") {
@@ -220,16 +234,46 @@ if ($tee == 'LISAA') {
 					kentta04    = '$paivarahat',
 					tapa 		= '$tapa',
 					tyyppi 		= 'kalenteri'";
-		mysql_query($query) or pupe_error($query);
+		pupe_query($query);
+
+		$uusi_tunnus = mysql_insert_id();
+
+		// P‰ivitet‰‰n liitetiedosto uudelle tunnukselle,
+		// mik‰li aikaisempi kalenterimerkint‰ p‰ivittyi.
+		// T‰m‰ siis sen takia, ett‰ kalenterimerkinn‰n p‰ivitys
+		// tapahtuu poistamalla vanha merkint‰ ja lis‰‰m‰ll‰ uusi p‰ivitetty
+		// versio merkinn‰st‰.
+		if ($tunnus > 0 and $uusi_tunnus > 0) {
+			$query = "  UPDATE liitetiedostot
+						SET liitostunnus = $uusi_tunnus
+						WHERE liitostunnus = $tunnus
+						AND liitos = 'kalenterimerkint‰'
+						AND yhtio = '{$kukarow['yhtio']}'";
+			pupe_query($query);
+		}
+
+		// Tallenna liitetiedosto formista
+		if (is_uploaded_file($_FILES['liitetiedosto']['tmp_name']) === TRUE) {
+			tallenna_liite('liitetiedosto', 'kalenterimerkint‰', $uusi_tunnus, '');
+		}
 	}
 }
 
 // ollaan painettu poista nappia
-if ($tee == "POISTA") {
-	$query ="	DELETE FROM kalenteri
-				WHERE tunnus = '$tunnus'
+if ($tee == "POISTA" and (int) $tunnus > 0) {
+	$tunnus = (int) $tunnus;
+
+	$query = "	DELETE FROM kalenteri
+				WHERE tunnus = $tunnus
 				$konsernit";
-	mysql_query($query) or pupe_error($query);
+	pupe_query($query);
+
+	// Poistetaan myˆs liitetiedostot t‰lle kalenterimerkinn‰lle
+	$query = "  DELETE FROM liitetiedostot
+				WHERE liitostunnus = {$tunnus}
+				AND liitos = 'kalenterimerkint‰'
+				AND yhtio = '{$kukarow['yhtio']}'";
+	pupe_query($query);
 }
 
 // tehd‰‰n lis‰ys ruutu ja laitetaan kaikki muuttujaan jotta voidaan echota sit oikeessa kohdassa
@@ -250,8 +294,8 @@ if($tee == "SYOTA") {
 					$konsernit
 					$kukalisa
 					and tyyppi = 'kalenteri'";
-		$res  = mysql_query($query) or pupe_error($query);
-		$irow = mysql_fetch_array($res);
+		$res  = pupe_query($query);
+		$irow = mysql_fetch_assoc($res);
 
 		$viesti 	= $irow["kentta01"];
 		$kilometrit	= $irow["kentta03"];
@@ -283,7 +327,7 @@ if($tee == "SYOTA") {
 	$lisayskello = $kello;
 
 	$lisays =  "
-		<td colspan='10'><form method='POST'>
+		<td colspan='10'><form method='POST' enctype='multipart/form-data'>
 		<input type='hidden' name='tee' value='LISAA'>
 		<input type='hidden' name='lopetus' value='$lopetus'>
 		<input type='hidden' name='valitut' value='".urlencode($valitut)."'>
@@ -352,7 +396,7 @@ if($tee == "SYOTA") {
 
 	$lisays .= "<tr><td>".t("Tapa").":</td><td><select name='tapa'>";
 
-	while ($vrow=mysql_fetch_array($vresult)) {
+	while ($vrow=mysql_fetch_assoc($vresult)) {
 		$sel="";
  		if ($tapa == $vrow["selitetark"]) {
 			$sel = "selected";
@@ -374,9 +418,9 @@ if($tee == "SYOTA") {
 					and oikeu.nimi		= 'crm/kalenteri.php'
 					and kuka.tunnus <> '$kukarow[tunnus]'
 					ORDER BY kuka.nimi";
-		$result = mysql_query($query) or pupe_error($query);
+		$result = pupe_query($query);
 
-		while ($row = mysql_fetch_array($result)) {
+		while ($row = mysql_fetch_assoc($result)) {
 			$sel = $kenelle == $row['kuka'] ? ' selected' : '';
 			$lisays .= "<option value='$row[kuka]'$sel>$row[nimi]</option>";
 		}
@@ -407,6 +451,33 @@ if($tee == "SYOTA") {
 	$lisays .= "<tr><td valign='top'>".t("P‰iv‰rahat").":</td>";
 	$lisays .= "<td>
 		  		<input name='paivarahat' value='$paivarahat'><br>
+		  		</td>
+		  		</tr>";
+
+	$query = "	SELECT *
+				from liitetiedostot
+				where yhtio = '$kukarow[yhtio]'
+				and liitos  = 'kalenterimerkint‰'
+				and liitostunnus = '$tunnus'";
+	$liiteres = pupe_query($query);
+
+	if (mysql_num_rows($liiteres) > 0) {
+		$lisays .= "<tr><td valign='top'>".t("Liitetiedostot").":</td>";
+		$lisays .= "<td><table><tbody>";
+
+		while ($liiterow = mysql_fetch_assoc($liiteres)) {
+			$lisays .= '<tr>';
+			$lisays .= "<td><a href=\"#\" onclick=\"window.open('".$palvelin2."view.php?id=$liiterow[tunnus]', '_blank' ,'toolbar=0,scrollbars=1,location=0,statusbar=0,menubar=0,resizable=1,left=200,top=100,width=800,height=600'); return false;\">$liiterow[filename]</a></td>";
+			$lisays .= '<td><a onclick="if (window.confirm(\'Haluatko poistaa liitteen '. $liiterow['filename'] .'\')===false ) return false;" href="'. $_SERVER['REQUEST_URI'] .'&poista_liite='. $liiterow['tunnus'] .'">Poista</a></td>';
+			$lisays .= '</tr>';
+		}
+
+		$lisays .= "</tbody></table></td></tr>";
+	}
+
+	$lisays .= "<tr><td valign='top'>".t("Lis‰‰ uusi Liitetiedosto").":</td>";
+	$lisays .= "<td>
+		  		<input name='liitetiedosto' type='file'><br>
 		  		</td>
 		  		</tr>";
 
@@ -446,7 +517,7 @@ if ($toim != 'TYOMAARAYS_ASENTAJA') {
 				and kalenteri.kuittaus='K'
 				$konsernit
 				ORDER BY kalenteri.pvmalku desc";
-	$result = mysql_query($query) or pupe_error($query);
+	$result = pupe_query($query);
 
 	if (mysql_num_rows($result) > 0) {
 	        echo "<table width='100%'>";
@@ -455,7 +526,7 @@ if ($toim != 'TYOMAARAYS_ASENTAJA') {
 	        echo "</tr>";
 
 
-	        while ($prow = mysql_fetch_array ($result)) {
+	        while ($prow = mysql_fetch_assoc ($result)) {
 	                echo "	<form action='kuittaamattomat.php?tee=A&kaletunnus=$prow[tunnus]&kuka=$prow[kuka]' method='post'>
 							<input type='hidden' name='lopetus' value='$lopetus'>
 							<tr>";
@@ -533,7 +604,7 @@ for ($i=1; $i <= days_in_month($kuu, $year); $i++) {
 				and kuka in ($vertaa)
 				$konsernit
 				and tyyppi = 'kalenteri'";
-	$result = mysql_query($query) or pupe_error($query);
+	$result = pupe_query($query);
 
 	//v‰ritet‰‰n t‰m‰n p‰iv‰n pvm omalla v‰rill‰...
 	if ((date("j")==$i) and (date("n")==$kuu) and (date("Y")==$year)) {
@@ -615,9 +686,9 @@ if ($yhtiorow["monikayttajakalenteri"] == "" or $kukarow["asema"] == "MP") {
 				and oikeu.nimi		= 'crm/kalenteri.php'
 				and kuka.tunnus <> '$kukarow[tunnus]'
 				ORDER BY kuka.nimi";
-	$result = mysql_query($query) or pupe_error($query);
+	$result = pupe_query($query);
 
-	while ($row = mysql_fetch_array($result)) {
+	while ($row = mysql_fetch_assoc($result)) {
 		$checked = '';
 		if (in_array("$row[kuka]", $ruksatut)) {
 			$checked = 'checked';
@@ -646,7 +717,7 @@ $query = "	SELECT kalenteri.asiakas, kalenteri.liitostunnus, kentta01, tapa, kuk
 			and kokopaiva != ''
 			$konsernit
 			order by pvmalku";
-$result = mysql_query($query) or pupe_error($query);
+$result = pupe_query($query);
 
 if (mysql_num_rows($result) > 0) {
 	echo "<table width='100%'>";
@@ -654,15 +725,15 @@ if (mysql_num_rows($result) > 0) {
 	echo "<th colspan='4'>".t("Kokop‰iv‰n tapahtumat")."</th>";
 	echo "</tr>";
 
-	while ($prow = mysql_fetch_array ($result)) {
+	while ($prow = mysql_fetch_assoc ($result)) {
 
 		//haetaan asiakkaan tiedot
 		$query = "	SELECT *
 					from asiakas
 					where yhtio = '$prow[yhtio]'
 					and tunnus  = '$prow[liitostunnus]'";
-		$asres = mysql_query($query) or pupe_error($query);
-		$asiak = mysql_fetch_array($asres);
+		$asres = pupe_query($query);
+		$asiak = mysql_fetch_assoc($asres);
 
 		echo "<tr>";
 		echo "<td>$prow[tapa]</td>";
@@ -752,9 +823,9 @@ while ($date != $whileloppu) {
 				and pvmalku  <= '$year-$kuu-$paiva $date:00'
 				and pvmloppu >= '$year-$kuu-$paiva $date:00'
 				and kokopaiva = ''";
-	$result = mysql_query($query) or pupe_error($query);
+	$result = pupe_query($query);
 
-	while ($row=mysql_fetch_array($result)) {
+	while ($row=mysql_fetch_assoc($result)) {
 		//tutkitaan kuinka monta p‰‰llekk‰ist‰ on parhaimmillaan...
 		if (mysql_num_rows($result)>$max) {
 			$max = mysql_num_rows($result);
@@ -797,7 +868,7 @@ while ($kello_nyt != $whileloppu) {
 				$konsernit
 				HAVING pvmalku  = '$year-$kuu-$paiva $kello_nyt:00'
 				order by kesto desc, pvmalku";
-	$result = mysql_query($query) or pupe_error($query);
+	$result = pupe_query($query);
 
 	//lasketaan montako p‰‰llekk‰ist‰ on t‰h‰n kellonaikaan
 	$aquery = "	SELECT
@@ -813,7 +884,7 @@ while ($kello_nyt != $whileloppu) {
 				pvmalku  <= '$year-$kuu-$paiva $kello_nyt:00'
 				and pvmalku > '$year-$kuu-$paiva 00:00:00'
 				and pvmloppu >  '$year-$kuu-$paiva $kello_nyt:00'";
-	$aresult = mysql_query($aquery) or pupe_error($aquery);
+	$aresult = pupe_query($aquery);
 
 	$nyt = mysql_num_rows($aresult);
 	$tyhjaa = $max-$nyt;
@@ -829,7 +900,7 @@ while ($kello_nyt != $whileloppu) {
 
 		if (mysql_num_rows($result) > 0) {
 
-			while ($row = mysql_fetch_array($result)) {
+			while ($row = mysql_fetch_assoc($result)) {
 
 				$kesto=round(($row['kesto']/60)/30); //kuinka monta solua t‰m‰ itemi kest‰‰
 
@@ -838,8 +909,8 @@ while ($kello_nyt != $whileloppu) {
 					$query = "	SELECT *
 								from asiakas
 								where tunnus = '$row[liitostunnus]' ".str_ireplace("kalenteri.", "", $konsernit);
-					$asres = mysql_query($query) or pupe_error($query);
-					$asiak = mysql_fetch_array($asres);
+					$asres = pupe_query($query);
+					$asiak = mysql_fetch_assoc($asres);
 				}
 
 				if ($kons == 1) {
@@ -870,8 +941,8 @@ while ($kello_nyt != $whileloppu) {
 								FROM avainsana
 								WHERE laji = 'KALETAPA' and selitetark='$row[tapa]' ".str_ireplace("kalenteri.", "", $konsernit)."
 								ORDER BY selite+0, laji, jarjestys, selite";
-					$varires = mysql_query($query) or pupe_error($query);
-					$varirow = mysql_fetch_array($varires);
+					$varires = pupe_query($query);
+					$varirow = mysql_fetch_assoc($varires);
 
 					if ($varirow["selitetark_2"] != "") {
 						$varilisa = "background-color: $varirow[selitetark_2];";
@@ -898,7 +969,24 @@ while ($kello_nyt != $whileloppu) {
 					}
 				}
 
-				echo "<br>$row[kentta01]</td>";
+				echo "<br>$row[kentta01]<br />";
+
+				$query = "	SELECT *
+							from liitetiedostot
+							where yhtio = '$kukarow[yhtio]'
+							and liitos  = 'kalenterimerkint‰'
+							and liitostunnus = '$row[tunnus]'";
+				$liiteres = pupe_query($query);
+
+				if (mysql_num_rows($liiteres) > 0) {
+					echo "<br />";
+
+					while ($liiterow = mysql_fetch_assoc($liiteres)) {
+						echo "<a href=\"#\" onclick=\"window.open('".$palvelin2."view.php?id=$liiterow[tunnus]', '_blank' ,'toolbar=0,scrollbars=1,location=0,statusbar=0,menubar=0,resizable=1,left=200,top=100,width=800,height=600'); return false;\">$liiterow[filename]</a>&nbsp;";
+					}
+				}
+
+				echo "</td>";
 			}
 
 			if ($tyhjaa > 0) {
@@ -924,5 +1012,3 @@ echo "</tr>";
 echo "</table>";
 
 require ("inc/footer.inc");
-
-?>
